+++ contentcuration/contentcuration/frontend/channelEdit/index.js
@@ -8,17 +8,3 @@ startApp({
   router,
   index: ChannelEditIndex,
 });
-
-// // TODO: Remove block once integrated into tree
-// window.preferences = {
-//   license: 'Public Domain',
-//   copyright_holder: 'Default Copyright Holder',
-// };
-// State.currentNode = {
-//   id: window.root_id,
-//   title: 'Sandbox Topic',
-//   metadata: {
-//     max_sort_order: 0,
-//   },
-// };
-// // End remove block
