+++ contentcuration/contentcuration/frontend/shared/constants.js
@@ -39,3 +39,17 @@ export const kindToIconMap = {
   topic: 'folder',
   video: 'theaters',
 };
+
+// File constants
+export const fileErrors = {
+  NO_STORAGE: 'NO_STORAGE',
+  WRONG_TYPE: 'WRONG_TYPE',
+  TOO_LARGE: 'TOO_LARGE',
+  UPLOAD_FAILED: 'UPLOAD_FAILED',
+  URL_EXPIRED: 'URL_EXPIRED',
+  CHECKSUM_HASH_FAILED: 'CHECKSUM_HASH_FAILED',
+};
+
+export const MAX_FILE_SIZE = 209715200;
+
+export const ASPECT_RATIO = 16 / 9;
