--- contentcuration/contentcuration/frontend/shared/views/files/constants.js
@@ -1,12 +1,0 @@
-// File constants
-export const fileErrors = {
-  NO_STORAGE: 'NO_STORAGE',
-  WRONG_TYPE: 'WRONG_TYPE',
-  TOO_LARGE: 'TOO_LARGE',
-  UPLOAD_FAILED: 'UPLOAD_FAILED',
-  URL_EXPIRED: 'URL_EXPIRED',
-};
-
-export const MAX_FILE_SIZE = 209715200;
-
-export const ASPECT_RATIO = 16 / 9;
